# services/theatre/theatre-spec/src/main/resources/openapi.yaml
openapi: 3.0.0
info:
  title: bookmy
  version: '1.0'
servers:
  - url: 'http://localhost:3000'
paths:
  '/theatres/shows/{movieName}':
    parameters:
      - schema:
          type: string
        name: movieName
        in: path
        required: true
    get:
      summary: Fetch shows by theatre
      tags:
        - shows
      responses:
        '200':
          $ref: '#/components/responses/TheatresGetResponseBody'
      operationId: getShowsByTheatre
      description: Get theatres with shows for given movie name on given date.
      parameters:
        - schema:
            type: string
            format: date
          in: query
          name: date
          description: date of show
        - schema:
            type: string
          in: query
          name: city
          description: city in which you want to watch show
  /theatres/shows:
    parameters: []
    post:
      tags:
        - shows
      summary: Create Show
      operationId: createShows
      responses:
        '200':
          $ref: '#/components/responses/TheatresGetResponseBody'
      description: ''
      requestBody:
        $ref: '#/components/requestBodies/TheatrePostRequestBody'
    put:
      tags:
        - shows
      summary: ''
      operationId: updateShow
      responses:
        '200':
          description: OK
      description: ''
      requestBody:
        $ref: '#/components/requestBodies/TheatrePostRequestBody'
    delete:
      tags:
        - shows
      summary: ''
      operationId: deleteShow
      responses:
        '200':
          description: OK
      parameters:
        - schema:
            type: integer
          in: query
          name: showId
          description: show id to delete
components:
  schemas:
    Theatre:
      type: object
      properties:
        theatreId:
          type: integer
        name:
          type: string
        city:
          type: string
        movies:
          type: array
          items:
            $ref: '#/components/schemas/Movie'
    Movie:
      type: object
      properties:
        movieId:
          type: integer
        name:
          type: string
        releaseDate:
          type: string
          format: date
        shows:
          type: array
          items:
            $ref: '#/components/schemas/Show'
    Show:
      type: object
      properties:
        showId:
          type: integer
        showTime:
          type: string
          format: time
        showDate:
          type: string
          format: date
        language:
          type: string
        seatingCapacity:
          type: integer
        showStatus:
          type: string
          enum:
            - ON_TIME
            - CANCELLED
            - DELAYED
            - BOOKING_FREEZED
        houseFull:
          type: boolean
  responses:
    TheatresGetResponseBody:
      description: Example response
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: '#/components/schemas/Theatre'
  requestBodies:
    TheatrePostRequestBody:
      content:
        application/json:
          schema:
            type: object
            properties:
              show:
                $ref: '#/components/schemas/Show'
              theatreId:
                type: integer
              movieId:
                type: integer
            required:
              - theatreId
              - movieId